Refactored record detail
